Add files via upload
The current project elucidates a sales dashboard of a Supermarket shop for the years 2021 and 2022. This excel dashboard provides with a flexibility of interacting with various buttons that helps in gaining a deeper understanding on the sales based on Sales type, Payment mode, and months in a year. Key Performance Indicators like Total sales, total profits and Profit margin help in making important strategic decisions.